Conflicts:
	Web/Controllers/WebRequest_PropertyController.cs

+++ Business/AppAdvertorialModel.cs
@@ -24,9 +24,9 @@ namespace Business
         /// </summary>
         /// <param name="AccountMainID"></param>
         /// <returns></returns>
-        public IQueryable<AppAdvertorial> GetList(int AccountMainID, int AdverTorialType)
+        public IQueryable<AppAdvertorial> GetList(int AccountMainID, int AdverTorialType, int EnumAdverClass)
         {
-            var appadverlist = List().Where(a => a.AccountMainID == AccountMainID && a.EnumAdvertorialUType == AdverTorialType).OrderByDescending(a => a.stick).ThenByDescending(a => a.Sort).ThenByDescending(a => a.IssueDate);
+            var appadverlist = List().Where(a => a.AccountMainID == AccountMainID && a.EnumAdvertorialUType == AdverTorialType && a.EnumAdverClass == EnumAdverClass).OrderByDescending(a => a.stick).ThenByDescending(a => a.Sort).ThenByDescending(a => a.IssueDate);
             return appadverlist;
         }
 
@@ -62,26 +62,39 @@ namespace Business
                 var imageminiPath = string.Format("{0}\\{1}", accountPath, imageminiName);
                 var imageshowName = string.Format("{0}_{1}_{2}", token, "show", LastName);
                 var imageshowPath = string.Format("{0}\\{1}", accountPath, imageshowName);
+                
                 var lsImgPath = appadvertorial.MainImagPath;
-                var lsImaFilePath = HttpContext.Current.Server.MapPath(lsImgPath);
-                Tool.SuperGetPicThumbnail(lsImaFilePath, imagePath, 70, 640, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
-
-
-                Tool.SuperGetPicThumbnailJT(lsImaFilePath, imagePath2, 70, w, h, x1, y1, tw, th, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
-
-                Tool.SuperGetPicThumbnail(imagePath2, imageshowPath, 70, 480, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
-
-                if (File.Exists(imagePath2))
+                if (!string.IsNullOrEmpty(lsImgPath))
                 {
-                    File.Delete(imagePath2);
+                    var lsImaFilePath = HttpContext.Current.Server.MapPath(lsImgPath);
+                    Tool.SuperGetPicThumbnail(lsImaFilePath, imagePath, 70, 640, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
+
+
+                    Tool.SuperGetPicThumbnailJT(lsImaFilePath, imagePath2, 70, w, h, x1, y1, tw, th, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
+
+                    Tool.SuperGetPicThumbnail(imagePath2, imageshowPath, 70, 480, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
+
+                    if (File.Exists(imagePath2))
+                    {
+                        File.Delete(imagePath2);
+                    }
+
+                    //缩略图mini
+                    Tool.SuperGetPicThumbnail(imageshowPath, imageminiPath, 70, 200, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
+
+                    appadvertorial.MainImagPath = path + imageName;
+                    appadvertorial.AppShowImagePath = path + imageshowName;
+                    appadvertorial.MinImagePath = path + imageminiName;
                 }
-
-                //缩略图mini
-                Tool.SuperGetPicThumbnail(imageshowPath, imageminiPath, 70, 200, 0, System.Drawing.Drawing2D.SmoothingMode.HighQuality, System.Drawing.Drawing2D.CompositingQuality.HighQuality, System.Drawing.Drawing2D.InterpolationMode.HighQualityBicubic);
-
-                appadvertorial.MainImagPath = path + imageName;
-                appadvertorial.AppShowImagePath = path + imageshowName;
-                appadvertorial.MinImagePath = path + imageminiName;
+                else
+                {
+                    if (appadvertorial.EnumAdverClass == (int)EnumAdverClass.Advertising)
+                    {
+                        appadvertorial.MainImagPath = "~/Images/Advertising.jpg";
+                        appadvertorial.AppShowImagePath = "~/Images/Advertising.jpg";
+                        appadvertorial.MinImagePath = "~/Images/Advertising_show.jpg";
+                    }
+                }
             }
             catch (Exception ex)
             {
@@ -109,7 +122,7 @@ namespace Business
 
             if (appadvertorial.stick == 1)
             {
-                int cnt = EditAppAdvertorialStick(appadvertorial.ID, appadvertorial.stick, appadvertorial.AccountMainID, appadvertorial.Sort, appadvertorial.EnumAdvertorialUType);
+                int cnt = EditAppAdvertorialStick(appadvertorial.ID, appadvertorial.stick, appadvertorial.AccountMainID, appadvertorial.Sort, appadvertorial.EnumAdvertorialUType, appadvertorial.EnumAdverClass);
                 if (cnt <= 0)
                 {
                     result.HasError = true;
@@ -257,7 +270,7 @@ namespace Business
             if (appadvertorial.stick != appadvertorials.stick)
             {
 
-                int cnt = EditAppAdvertorialStick(appadvertorial.ID, appadvertorial.stick, appadvertorial.AccountMainID, appadvertorial.Sort, appadvertorial.EnumAdvertorialUType);
+                int cnt = EditAppAdvertorialStick(appadvertorial.ID, appadvertorial.stick, appadvertorial.AccountMainID, appadvertorial.Sort, appadvertorial.EnumAdvertorialUType, appadvertorial.EnumAdverClass);
                 if (cnt <= 0)
                 {
                     result.HasError = true;
@@ -282,7 +295,7 @@ namespace Business
 
         //修改置顶 isok 1 置顶 0 取消
         [Transaction]
-        public int EditAppAdvertorialStick(int ID, int isok, int accoutMainID, int Sort, int AdverTorialType)
+        public int EditAppAdvertorialStick(int ID, int isok, int accoutMainID, int Sort, int AdverTorialType, int EnumAdverClass)
         {
             string sql = "";
             if (isok == 1)
@@ -294,7 +307,7 @@ namespace Business
             else
             {
                 sql = string.Format("update AppAdvertorial set stick = {0}, sort=0 where id ={1}", isok, ID);
-                string sql2 = string.Format("update AppAdvertorial set Sort = (Sort-1) where stick = 1 and Sort>{0} and accountMainID={1} and EnumAdvertorialUType={2}", Sort, accoutMainID, AdverTorialType);
+                string sql2 = string.Format("update AppAdvertorial set Sort = (Sort-1) where stick = 1 and Sort>{0} and accountMainID={1} and EnumAdvertorialUType={2} and EnumAdverClass={3}", Sort, accoutMainID, AdverTorialType, EnumAdverClass);
                 base.SqlExecute(sql2);
             }
 
@@ -302,7 +315,7 @@ namespace Business
         }
 
         //排序 type 1 向上 0 向下
-        public int EditAppAdvertorialSort(int ID, int AccountMainID, int Sort, int type, int AdverTorialType)
+        public int EditAppAdvertorialSort(int ID, int AccountMainID, int Sort, int type, int AdverTorialType, int EnumAdverClass)
         {
             int cnt = List().Where(a => a.AccountMainID == AccountMainID && a.stick == 1).Count();
             if ((Sort == 1 && type == 0) || (Sort == cnt && type == 1))
@@ -314,13 +327,13 @@ namespace Business
                 string sql = "";
                 if (type == 1) //向上
                 {
-                    sql = string.Format(@"update AppAdvertorial set Sort = (Sort-1) where accountMainID={1}  and EnumAdvertorialUType={3} and Sort=({2}+1) and stick=1 
-                                    update AppAdvertorial set Sort = (Sort+1) where ID={0} ", ID, AccountMainID, Sort, AdverTorialType);
+                    sql = string.Format(@"update AppAdvertorial set Sort = (Sort-1) where accountMainID={1}  and EnumAdvertorialUType={3} and Sort=({2}+1) and stick=1 and EnumAdverClass={4} 
+                                    update AppAdvertorial set Sort = (Sort+1) where ID={0} ", ID, AccountMainID, Sort, AdverTorialType, EnumAdverClass);
                 }
                 else
                 {
-                    sql = string.Format(@"update AppAdvertorial set Sort = (Sort+1) where accountMainID={1}  and EnumAdvertorialUType={3} and Sort=({2}-1) and stick=1 
-                                    update AppAdvertorial set Sort = (Sort-1) where ID={0} ", ID, AccountMainID, Sort, AdverTorialType);
+                    sql = string.Format(@"update AppAdvertorial set Sort = (Sort+1) where accountMainID={1}  and EnumAdvertorialUType={3} and Sort=({2}-1) and stick=1 and EnumAdverClass={4} 
+                                    update AppAdvertorial set Sort = (Sort-1) where ID={0} ", ID, AccountMainID, Sort, AdverTorialType, EnumAdverClass);
                 }
                 return base.SqlExecute(sql);
             }
